Fixed code to pass more tests
git-svn-id: https://svn.apache.org/repos/asf/incubator/lucene.net/branches/Lucene.Net.3.0.3@1202583 13f79535-47bb-0310-9956-ffa450edef68

+++ trunk/test/core/Util/TestNumericUtils.cs
@@ -79,9 +79,9 @@ namespace Lucene.Net.Util
                 neededBounds.MoveNext();
                 unchecked
                 {
-                    Assert.AreEqual((long)neededBounds.Current, (ulong) Number.URShift(min, shift), "inner min bound");
+                    Assert.AreEqual((long)neededBounds.Current, Number.URShift(min, shift), "inner min bound");
                     neededBounds.MoveNext();
-                    Assert.AreEqual((long)neededBounds.Current, (ulong)Number.URShift(max, shift), "inner max bound");
+                    Assert.AreEqual((long)neededBounds.Current, Number.URShift(max, shift), "inner max bound");
                 }
 			}
 		}
